Split python module internals
