+++ crates/djls-project/src/lib.rs
@@ -18,6 +18,8 @@ pub use python::Interpreter;
 pub use python::InvalidModuleName;
 pub use python::PythonModule;
 pub use python::PythonModuleName;
+pub use python::SearchPath;
+pub use python::SearchPaths;
 pub use refresh::RefreshCountUnits;
 pub use refresh::RefreshData;
 pub use refresh::RefreshPart;
@@ -26,8 +28,6 @@ pub use refresh::RefreshTaskDescriptor;
 pub use refresh::RefreshTaskGroup;
 pub use refresh::apply_refresh;
 pub use refresh::refresh_tasks;
-pub use resolve::SearchPath;
-pub use resolve::SearchPaths;
 pub use resolve::model_modules;
 pub use templates::ArgumentCountConstraint;
 pub use templates::AsVar;
